@@ -1,9 +1,7 @@
 <?php
 
 namespace App\Http\Controllers\Auth;
-//use  recaptchalib;
-use App\Recaptchalib;
-//require ('app/Recaptchalib.php');
+
 use App\Http\Controllers\Controller;
 use App\Providers\RouteServiceProvider;
 use Illuminate\Foundation\Auth\AuthenticatesUsers;
@@ -28,8 +26,7 @@ class LoginController extends Controller
      *
      * @var string
      */
-    //protected $redirectTo = RouteServiceProvider::HOME;
-    protected $redirectTo = '/home';
+    protected $redirectTo = RouteServiceProvider::HOME;
 
     /**
      * Create a new controller instance.
