@@ -2,14 +2,9 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Contracts\Support\Renderable;
+
+use App\Models\Post;
 use Illuminate\Http\Request;
-use App\Image;
-use Illuminate\Support\Facades\Storage;
-use Illuminate\Support\Facades\File;
-
-
-
 
 class HomeController extends Controller
 {
@@ -20,26 +15,18 @@ class HomeController extends Controller
      */
     public function __construct()
     {
-
         $this->middleware('auth');
     }
 
     /**
      * Show the application dashboard.
      *
-     * @return Renderable
+     * @return \Illuminate\Contracts\Support\Renderable
      */
     public function index()
     {
-        $images = Image::orderBy('id_image', 'desc')->paginate(5);
-
-        //var_dump($images);
-
-        return view('home')->with(
-            'images', $images
-        );
-
+        $images = Post::orderBy('id', 'desc')->paginate(10);
+        return view('home')
+            ->with('images', $images);
     }
-
-
 }
