@@ -3,14 +3,13 @@
 namespace App\Http\Middleware;
 
 use Illuminate\Auth\Middleware\Authenticate as Middleware;
-use Illuminate\Http\Request;
 
 class Authenticate extends Middleware
 {
     /**
      * Get the path the user should be redirected to when they are not authenticated.
      *
-     * @param  Request  $request
+     * @param  \Illuminate\Http\Request  $request
      * @return string|null
      */
     protected function redirectTo($request)
