@@ -1,65 +1,65 @@
-
-
 @extends('layouts.app')
 <title>Login</title>
 @section('content')
-<head>
-    <script src="https://www.google.com/recaptcha/api.js?hl=es" async defer></script>
-    <link rel="preconnect" href="https://www.google.com">
-    <link rel="preconnect" href="https://www.gstatic.com" crossorigin>
-    <!-- Google Font: Source Sans Pro -->
-    <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Source+Sans+Pro:300,400,400i,700&display=fallback">
-    <!-- Font Awesome -->
-    <link rel="stylesheet" href="./plugins/fontawesome-free/css/all.min.css">
-    <!-- icheck bootstrap -->
-    <link rel="stylesheet" href="./plugins/icheck-bootstrap/icheck-bootstrap.min.css">
-    <!-- Theme style -->
-    <link rel="stylesheet" href="./css/adminlte.min.css">
-    {!! RecaptchaV3::initJs() !!}
-</head>
-<body>
-<script async src="https://www.google.com/recaptcha/api.js"></script>
+    <head>
+        <script src="https://www.google.com/recaptcha/api.js?hl=es" async defer></script>
+        <link rel="preconnect" href="https://www.google.com">
+        <link rel="preconnect" href="https://www.gstatic.com" crossorigin>
+        <!-- Google Font: Source Sans Pro -->
+        <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Source+Sans+Pro:300,400,400i,700&display=fallback">
+        <!-- Font Awesome -->
+        <link rel="stylesheet" href="./plugins/fontawesome-free/css/all.min.css">
+        <!-- icheck bootstrap -->
+        <link rel="stylesheet" href="./plugins/icheck-bootstrap/icheck-bootstrap.min.css">
+        <!-- Theme style -->
+        <link rel="stylesheet" href="./css/adminlte.min.css">
+        {!! RecaptchaV3::initJs() !!}
 
-<div class="hold-transition login-page">
-<div class="login-box">
-    <!-- /.login-logo -->
-    <div class="card text-white bg-dark mb-3">
-        <div class="card-header text-center">
-            <a href="#" class="h1"><b>SAIEU</b></a>
-        </div>
+
+    </head>
+    <body>
+    <script async src="https://www.google.com/recaptcha/api.js"></script>
+    <div class="hold-transition login-page">
+        <div class="login-box">
+            <!-- /.login-logo -->
+            <div class="card text-white bg-dark mb-3">
+                <div class="card-header text-center">
+                    <a href="#" class="h1"><b>SAIEU</b></a>
+                </div>
                 <div class="card-body">
                     <p class="login-box-msg">Ingresa con tus credenciales para iniciar sesión</p>
+
                     <form method="POST" action="{{ route('login') }}">
                         @csrf
                         <div class="input-group mb-3">
-                                <input id="email" type="email" class="form-control @error('email') is-invalid @enderror" name="email" placeholder="Email" value="{{ old('email') }}" required autocomplete="email" autofocus>
+                            <input id="email" type="email" class="form-control @error('email') is-invalid @enderror" name="email" placeholder="Correo" value="{{ old('email') }}" required autocomplete="email" autofocus>
                             <div class="input-group-append">
                                 <div class="input-group-text">
                                     <span class="fas fa-envelope"></span>
                                 </div>
                             </div>
                         </div>
-                                @error('email')
-                                    <span class="invalid-feedback" role="alert">
+                        @error('email')
+                        <span class="invalid-feedback" role="alert">
                                         <strong>{{ $message }}</strong>
                                     </span>
 
-                                @enderror
+                        @enderror
 
-                            <div class="input-group mb-3">
-                                <input id="password" type="password" class="form-control @error('password') is-invalid @enderror" name="password" placeholder="Password" required autocomplete="current-password">
-                                <div class="input-group-append">
-                                    <div class="input-group-text">
-                                        <span class="fas fa-lock"></span>
-                                    </div>
+                        <div class="input-group mb-3">
+                            <input id="password" type="password" class="form-control @error('password') is-invalid @enderror" name="password" placeholder="Contraseña" required autocomplete="current-password">
+                            <div class="input-group-append">
+                                <div class="input-group-text">
+                                    <span class="fas fa-lock"></span>
                                 </div>
                             </div>
+                        </div>
 
-                                    @error('password')
-                                    <span class="invalid-feedback" role="alert">
+                        @error('password')
+                        <span class="invalid-feedback" role="alert">
                                         <strong>{{ $message }}</strong>
                                     </span>
-                                @enderror
+                        @enderror
 
 
 
@@ -82,46 +82,44 @@
                                         <strong>{{ $errors->first('g-recaptcha-response') }}</strong>
                                     </span>
                                 @endif
-                        <div class="form-group row mb-0">
-                            <div class="col-md-8 offset-md-4">
-                                <button class=" btn btn-primary "
+                                <div class="form-group row mb-0">
+                                    <div class="col-md-8 offset-md-4">
+                                        <button class=" btn btn-primary "
 
-                                         type="submit" >
-                                    <span class="fas fa-sign-in-alt"></span>
-                                    Inicia Sesión
-                                </button>
+                                                type="submit" >
+                                            <span class="fas fa-sign-in-alt"></span>
+                                            Inicia Sesión
+                                        </button>
+                                    </div>
+                                </div>
                             </div>
                         </div>
-                            </div>
-                        </div>
-                                </form>
-                                @if (Route::has('password.request'))
-                                    <a class="btn btn-link" href="{{ route('password.request') }}">
-                                        ¿Olvidaste tu contraseña?
-                                    </a>
-                                @endif
+                    </form>
+                    @if (Route::has('password.request'))
+                        <a class="btn btn-link" href="{{ route('password.request') }}">
+                            ¿Olvidaste tu contraseña?
+                        </a>
+                    @endif
                     <br>
                     @if (Route::has('register'))
 
-                            <a class="btn btn-link" href="{{ route('register') }}">
-                                ¿No tienes una cuenta? Registrate ahora </a>
+                        <a class="btn btn-link" href="{{ route('register') }}">
+                            ¿No tienes una cuenta? Registrate ahora </a>
 
                     @endif
 
-                            </div>
-                        </div>
-                    </div>
                 </div>
             </div>
-        @endsection
+        </div>
+    </div>
 
 
 
-<!-- jQuery -->
-<script src="./plugins/jquery/jquery.min.js"></script>
-<!-- Bootstrap 4 -->
-<script src="./plugins/bootstrap/js/bootstrap.bundle.min.js"></script>
-<!-- AdminLTE App -->
-<script src="./js/adminlte.min.js"></script>
+@endsection
+    <script src="./plugins/jquery/jquery.min.js"></script>
+    <!-- Bootstrap 4 -->
+    <script src="./plugins/bootstrap/js/bootstrap.bundle.min.js"></script>
+    <!-- AdminLTE App -->
+    <script src="./js/adminlte.min.js"></script>
 
-</body>
+    </body>
