@@ -1,29 +1,29 @@
 <!doctype html>
-<html lang="{{ str_replace('_', '-', app()->getLocale()) }}" xmlns="http://www.w3.org/1999/html">
+<html lang="{{ str_replace('_', '-', app()->getLocale()) }}">
 <head>
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1">
-<script>function darkmode() {
-        var cuerpoweb = document.body;
-        cuerpoweb.classList.toggle("oscuro");
-    }
+    <script>
 
-    tippy('#btn_users', {
-        content: "I'm a Tippy tooltip!",
-    });
+        function darkmode() {
+            var cuerpoweb = document.body;
+            cuerpoweb.classList.toggle("oscuro");
+        }
+
+        $(function () {
+            $('[data-toggle="tooltip"]').tooltip()
+        })
 
 
-</script>
+    </script>
     <!-- CSRF Token -->
     <meta name="csrf-token" content="{{ csrf_token() }}">
 
-    <title>{{ config('app.name', 'SAIEU-UT') }}</title>
+    <title>{{ config('app.name', 'Laravel') }}</title>
 
     <!-- Scripts -->
-
     <script src="{{ asset('js/app.js') }}" defer></script>
     <script src="{{ asset('js/main.js') }}"></script>
-
     <script src="https://unpkg.com/tippy.js@6/dist/tippy-bundle.umd.js"></script>
     <script src=" https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.11.0/umd/popper.min.js"></script>
     <script src="//cdn.jsdelivr.net/npm/sweetalert2@11"></script>
@@ -42,97 +42,97 @@
     @laravelPWA
 </head>
 <body>
-<div id="app">
-    <nav class="navbar navbar-expand-md navbar-light bg-success shadow-sm">
-        <div class="container">
-            <a class="navbar-brand" href="{{ route('home') }}">
+    <div id="app">
+        <nav class="navbar sticky-top navbar-expand-md navbar-light bg-success shadow-sm">
+            <div class="container-fluid">
+                <a class="navbar-brand" href="{{ url('home') }}">
+                    <img src="{{asset('images/png/SAIEU.png')}}" alt="Logo" width="60" height="40">
+                </a>
+                <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="{{ __('Toggle navigation') }}">
+                    <span class="navbar-toggler-icon"></span>
+                </button>
 
-                {{-- {{ config('app.name', 'Laravel') }} --}}
-                SAIEU-UT
-            </a>
-            <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="{{ __('Toggle navigation') }}">
-                <span class="navbar-toggler-icon"></span>
-            </button>
+                <div class="collapse navbar-collapse" id="navbarSupportedContent">
+                    <!-- Left Side Of Navbar -->
+                    <ul class="navbar-nav mr-auto">
 
-            <div class="collapse navbar-collapse" id="navbarSupportedContent">
-                <!-- Left Side Of Navbar -->
-                <ul class="navbar-nav mr-auto">
+                    </ul>
 
-                </ul>
-
-                <!-- Right Side Of Navbar -->
-                <ul class="navbar-nav ml-auto">
-                    <!-- Authentication Links -->
-                    @guest
-                        <li class="nav-item">
-                            <button type="button" class="btn btn-dark" onclick="darkmode()">Oscuro / Claro</button>
-
-                        </li>
-                        <li class="nav-item">
-                            <a class="nav-link" hidden href="{{ route('login') }}">Login</a>
-                        </li>
-                        @if (Route::has('register'))
+                    <!-- Right Side Of Navbar -->
+                    <ul class="navbar-nav ml-auto">
+                        <!-- Authentication Links -->
+                        @guest
                             <li class="nav-item">
-                                <a class="nav-link" hidden href="{{ route('register') }}">Registrarse</a>
+                                <button type="button" class="btn btn-dark" onclick="darkmode()">Oscuro / Claro</button>
+
                             </li>
-                        @endif
-                    @else
-                        <li class="nav-item">
-                            <button type="button" class="btn btn-dark" onclick="darkmode()">Oscuro / Claro</button>
 
-                        </li>
+                            @if (Route::has('login'))
+                                <li class="nav-item">
+                                    <a class="nav-link" href="{{ route('login') }}">{{ __('Login') }}</a>
+                                </li>
+                            @endif
 
-                        <li class="nav-item">
-                            <a href="{{route('user.index')}}" id="btn_users" class="nav-link"><i class="fas fa-users"></i></a>
-                        </li>
-                        <li class="nav-item">
-                            <a href="{{route('image.create')}}" id="btn_avisos" class="nav-link"><i class="fas fa-file-invoice"></i></a>
-                        </li>
-                        <li class="nav-item">
-                            <a href="{{route('print')}}" id="btn_pdf" class="nav-link"><i class="fas fa-file-pdf"></i></a>
-                        </li>
+                            @if (Route::has('register'))
+                                <li class="nav-item">
+                                    <a class="nav-link" href="{{ route('register') }}">{{ __('Register') }}</a>
+                                </li>
+                            @endif
+                        @else
+                            <li class="nav-item">
+                                <button type="button" class="btn btn-dark" onclick="darkmode()">Oscuro / Claro</button>
 
-                        <li class="nav-item dropdown">
-                            <a id="navbarDropdown" class="nav-link dropdown-toggle"  href="#" role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="true" v-pre>
-                                {{ Auth::user()->name }} <span class="caret"></span>
-                            </a>
+                            </li>
+
+                            <li class="nav-item">
+                                <a href="{{route('user.index')}}" id="btn_users" data-toggle="tooltip" data-placement="top" title="Usuarios" class="nav-link"><i class="fas fa-users"></i></a>
+                            </li>
+                            <li class="nav-item">
+                                <a href="{{route('image.create')}}" id="btn_avisos" data-toggle="tooltip" data-placement="top" title="Crear Aviso" class="nav-link"><i class="fas fa-file-invoice"></i></a>
+                            </li>
+                            <li class="nav-item">
+                                <a href="{{route('print')}}" id="btn_pdf" data-toggle="tooltip" data-placement="top" title="Generar reportes" target="_blank" class="nav-link"><i class="fas fa-file-pdf"></i></a>
+                            </li>
 
 
-                            <div class="dropdown-menu dropdown-menu-right" aria-labelledby="navbarDropdown">
-                                <a class="dropdown-item" href="{{action('UserController@profile',['id' => Auth::user()->id_user])}}">
+
+
+
+                            <li class="nav-item dropdown">
+                                <a id="navbarDropdown" class="nav-link dropdown-toggle" href="#" role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false" v-pre>
+                                    {{ Auth::user()->name }}
+                                </a>
+                                <div class="dropdown-menu dropdown-menu-right" aria-labelledby="navbarDropdown">
+                                <a class="dropdown-item" href="{{action('UserController@profile',['id' => Auth::user()->id])}}">
                                     Mis datos
                                 </a>
-                                <a class="dropdown-item" href="{{ route('config') }}">
+                                <a class="dropdown-item" href="{{ route('user.config') }}">
                                     Configuración
                                 </a>
 
+
 <center>
 
-
-                                <a class="dropdown-item" href="{{ route('logout') }}"
-                                   onclick="event.preventDefault();
+                                    <a class="dropdown-item" href="{{ route('logout') }}"
+                                       onclick="event.preventDefault();
                                                      document.getElementById('logout-form').submit();">
-                                    <span class="fas fa-power-off"></span>
-                                </a>
+                                        <span class="fas fa-power-off"></span>
+                                    </a>
 </center>
-
-                                <form id="logout-form" action="{{ route('logout') }}" method="POST" style="display: none;">
-                                    @csrf
-                                </form>
-                            </div>
-                        </li>
-                    @endguest
-                </ul>
+                                    <form id="logout-form" action="{{ route('logout') }}" method="POST" class="d-none">
+                                        @csrf
+                                    </form>
+                                </div>
+                            </li>
+                        @endguest
+                    </ul>
+                </div>
             </div>
-        </div>
-    </nav>
+        </nav>
 
-    <main class="py-4">
-        @yield('content')
-    </main>
-</div>
-<script>
-
-</script>
+        <main class="py-4">
+            @yield('content')
+        </main>
+    </div>
 </body>
 </html>
