@@ -3,14 +3,13 @@
 namespace Tests;
 
 use Illuminate\Contracts\Console\Kernel;
-use Illuminate\Foundation\Application;
 
 trait CreatesApplication
 {
     /**
      * Creates the application.
      *
-     * @return Application
+     * @return \Illuminate\Foundation\Application
      */
     public function createApplication()
     {
